refactor: split properties panel tests into separate spec files
Move PropertiesPanelHelper into propertiesPanel folder, split monolithic
spec into openClose, workflowOverview, nodeSelection, titleEditing,
searchFiltering, nodeSettings, infoTab, and errorsTab specs. Add
AGENTS.md with test structure documentation.

Fix color tests to use page.evaluate for node property verification
instead of brittle CSS background-color assertions.

Clean up redundant comments from PropertiesPanelHelper.

Addresses review feedback:
https://github.com/Comfy-Org/ComfyUI_frontend/pull/10548#discussion_r3005784036
https://github.com/Comfy-Org/ComfyUI_frontend/pull/10548#discussion_r3005784429
