+++ browser_tests/tests/propertiesPanel/PropertiesPanelHelper.ts
@@ -1,7 +1,7 @@
 import type { Locator, Page } from '@playwright/test'
 import { expect } from '@playwright/test'
 
-import { TestIds } from '../fixtures/selectors'
+import { TestIds } from '../../fixtures/selectors'
 
 export class PropertiesPanelHelper {
   readonly root: Locator
@@ -16,8 +16,7 @@ export class PropertiesPanelHelper {
     this.closeButton = this.root.locator('button[aria-pressed]')
   }
 
-  /** Tab navigation locators */
-  get tabs() {
+  get tabs(): Locator {
     return this.root.locator('nav button')
   }
 
@@ -25,7 +24,6 @@ export class PropertiesPanelHelper {
     return this.root.locator('nav button', { hasText: label })
   }
 
-  /** Title editing */
   get titleEditIcon(): Locator {
     return this.panelTitle.locator('i[class*="lucide--pencil"]')
   }
@@ -34,42 +32,34 @@ export class PropertiesPanelHelper {
     return this.root.getByTestId(TestIds.node.titleInput)
   }
 
-  /** Settings tab: node state buttons */
   getNodeStateButton(state: 'Normal' | 'Bypass' | 'Mute'): Locator {
     return this.root.locator('button', { hasText: state })
   }
 
-  /** Settings tab: color swatch by name */
   getColorSwatch(colorName: string): Locator {
     return this.root.locator(`[data-testid="${colorName}"]`)
   }
 
-  /** Settings tab: pinned toggle (PrimeVue ToggleSwitch wrapper div) */
   get pinnedSwitch(): Locator {
     return this.root.locator('[data-p-checked]').first()
   }
 
-  /** Subgraph edit button (gear icon in header) */
   get subgraphEditButton(): Locator {
     return this.root.locator('button:has(i[class*="lucide--settings-2"])')
   }
 
-  /** Panel content area */
   get contentArea(): Locator {
     return this.root.locator('.scrollbar-thin')
   }
 
-  /** Errors tab indicator icon */
   get errorsTabIcon(): Locator {
     return this.root.locator('nav i[class*="lucide--octagon-alert"]')
   }
 
-  /** Global settings: "View all settings" button */
   get viewAllSettingsButton(): Locator {
     return this.root.getByRole('button', { name: /view all settings/i })
   }
 
-  /** Collapse/expand toggle button */
   get collapseToggleButton(): Locator {
     return this.root.locator(
       'button:has(i[class*="lucide--chevrons-down-up"]), button:has(i[class*="lucide--chevrons-up-down"])'
